pyproject.toml
[tool.poetry]
name = "project"
version = "0.2.0"
description = ""
authors = ["CuasarN1 <55405975+CuasarN1@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.35"
fastapi = "^0.115.2"
pydantic-settings = "^2.5.2"
uvicorn = "^0.31.1"
asyncpg = "^0.29.0"
greenlet = "^3.1.1"
alembic = "^1.13.3"
tenacity = "^9.0.0"
email-validator = "^2.1.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
networkx = "^3.4.2"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
alembic = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
